# feat(server): improve get albums performance