--- a/server/apps/immich/src/api-v1/album/album.module.ts
+++ b/server/apps/immich/src/api-v1/album/album.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AlbumEntity, AssetEntity } from '@app/infra';
+import { AlbumEntity, AssetEntity, SharedLinkEntity } from '@app/infra';
 import { AlbumRepository, IAlbumRepository } from './album-repository';
 import { DownloadModule } from '../../modules/download/download.module';
 
@@ -12,7 +12,7 @@ const ALBUM_REPOSITORY_PROVIDER = {
 };
 
 @Module({
-  imports: [TypeOrmModule.forFeature([AlbumEntity, AssetEntity]), DownloadModule],
+  imports: [TypeOrmModule.forFeature([AlbumEntity, AssetEntity, SharedLinkEntity]), DownloadModule],
   controllers: [AlbumController],
   providers: [AlbumService, ALBUM_REPOSITORY_PROVIDER],
   exports: [ALBUM_REPOSITORY_PROVIDER],