@michelheusschen 2023-03-05
Will close this for now as maintainability is an issue. This does need improvement though and will get back to this later once I've reworked some things.